chore(repos): restructure folders
