+++ src/infra/persistence/repos/accounting-entity.repo.impl.ts
@@ -1,7 +1,7 @@
 import IAccountingEntityRepo from '../../../domain/accounting/repos/accounting-entity.repo';
 import { accountingEntitiesInCore } from '../drizzle/schema';
 import accountingEntityMapper from '../../../app/mappers/accounting-entity.mapper';
-import getDbQuery from './query';
+import getDbQuery from './helpers/query';
 
 const accountingEntityRepo: IAccountingEntityRepo = {
   save: async (domain, options) => {
