+++ src/infra/persistence/repos/currency.repo.impl.ts
@@ -2,7 +2,7 @@ import { eq } from 'drizzle-orm';
 import ICurrencyRepo from '../../../domain/currency/repos/currency.repo';
 import { currenciesInCore as currencies } from '../drizzle/schema';
 import currencyMapper from '../../../app/mappers/currency.mapper';
-import getDbQuery from './query';
+import getDbQuery from './helpers/query';
 
 const currencyRepo: ICurrencyRepo = {
   save: async (payload, options) => {
